Renaming App to app, and including prettier deps
