+++ package.json
@@ -22,7 +22,9 @@
 	"devDependencies": {
 		"@woocommerce/dependency-extraction-webpack-plugin": "1.4.0",
 		"@woocommerce/eslint-plugin": "1.1.0",
-		"@wordpress/scripts": "^13.0.3"
+		"@wordpress/prettier-config": "^1.0.1",
+		"@wordpress/scripts": "^13.0.3",
+		"prettier": "^2.2.1"
 	},
 	"dependencies": {
 		"@woocommerce/data": "^1.1.1",
