+++ src/index.js
@@ -6,7 +6,7 @@ import { render } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { App } from './App';
+import { App } from './app';
 import './index.scss';
 
 const appRoot = document.getElementById(
